servers/ga4_server/requirements.txt
mcp==1.9.0
pydantic-settings==2.9.1
python-dotenv==1.1.0
uvicorn==0.34.2
starlette==0.46.2
sse-starlette==2.3.5
requests==2.32.3
httpx==0.28.1
google-analytics-data==0.18.19
google-analytics-admin==0.13.0
annotated-types
anyio
certifi
click
h11
httpcore
httpx-sse
idna
pydantic
pydantic_core
python-multipart
sniffio
typing-inspection
typing_extensions 